fix intermediate mantissa being shifted out

+++ softfloat/s_roundPackToF32.c
@@ -54,7 +54,7 @@ float32_t
     *------------------------------------------------------------------------*/
     softfloat_intermediateResult.sign     = sign;
     softfloat_intermediateResult.exp      = exp;
-    softfloat_intermediateResult.sig64    = sig<<32;
+    softfloat_intermediateResult.sig64    = ((uint64_t)sig)<<32;
     softfloat_intermediateResult.sig0     = 0;
     softfloat_intermediateResult.sigExtra = 0;
